Begin splitting out controller
As part of #112

This also aids with #124
